Swap the colon and quote keys.
The colon key on the right pinkie home row is a total waste.
Furthermore, I'd like a not-that-often used key to the right of
the home row to use as a mod shift key. The quote key is way too
often used, i.e. wrapping a string in quotes.

This creates symmetrical left and right side shift alignment,
and adds back in right side ctrl/meta modifiers.