Swap keys for numpad layer and gaming layer.
Since the numpad layer only uses the left side, it doesn't make much sense
for the key to switch to it to be on the right side. This swaps that with
the key for the gaming layer, which happens less frequently, and is okay to
be on the right side.