Swap around the right quote/semi-colon.
I find even after months, that I'm still messing up the semicolon
and quote keys. This tries to fix that by moving the semicolon key
back to it's original spot under the right pinky, moving the quote
key to the backslash spot, and moving the backslash key to the
backspace spot.

This seems less error prone, since I don't oversubscribe the shift key,
the semi-colon is in it's muscle memory spot, the quote key is just shifted
slightly from its original position, and the backslash key isn't used that
often anyway.