Add up/down arrows to left side bottom keys.
Up/down are really useful to have, and switching to another layer to use them
is annoying. So I moved them to the bottom keys on the left side, and moved the
volume adjustment keys to the right side.

Generally, using more of my layer 0 keys is a good idea.